HSEARCH-4133 Move non-regression tests next to other, related tests
Instead of putting them in their own package.

This is so I can more easily isolate all tests related to automatic
indexing in the next few commits.
